Invite auto sends to chat. ONLY PRIVATE
- Tried to fix Messages css but not working
